chore: move tests to mentions
