+++ framework/core/tests/integration/api/discussions/CreateTest.php
@@ -72,97 +72,6 @@ class CreateTest extends TestCase
     /**
      * @test
      */
-    public function cannot_create_discussion_without_content_property()
-    {
-        $this->extend(
-            (new Extend\Formatter)
-                ->unparse(function ($context, string $content) {
-                    return $content;
-                }),
-            (new Extend\Event())
-                ->listen(\Flarum\Post\Event\Saving::class, function ($event) {
-                    $event->post->content;
-                })
-        );
-
-        $response = $this->send(
-            $this->request('POST', '/api/discussions', [
-                'authenticatedAs' => 1,
-                'json' => [
-                    'data' => [
-                        'attributes' => [
-                            'title' => 'Test post',
-                        ],
-                    ],
-                ],
-            ])
-        );
-
-        $this->assertEquals(422, $response->getStatusCode());
-
-        $body = (string) $response->getBody();
-        $this->assertJson($body);
-        $this->assertEquals([
-            'errors' => [
-                [
-                    'status' => '422',
-                    'code' => 'validation_error',
-                    'detail' => 'The content field is required.',
-                    'source' => ['pointer' => '/data/attributes/content'],
-                ],
-            ],
-        ], json_decode($body, true));
-    }
-
-    /**
-     * @test
-     */
-    public function cannot_create_discussion_with_content_set_to_null()
-    {
-        $this->extend(
-            (new Extend\Formatter)
-                ->unparse(function ($context, string $content) {
-                    return $content;
-                }),
-            (new Extend\Event())
-                ->listen(\Flarum\Post\Event\Saving::class, function ($event) {
-                    $event->post->content;
-                })
-        );
-
-        $response = $this->send(
-            $this->request('POST', '/api/discussions', [
-                'authenticatedAs' => 1,
-                'json' => [
-                    'data' => [
-                        'attributes' => [
-                            'title' => 'Test post',
-                            'content' => null,
-                        ],
-                    ],
-                ],
-            ])
-        );
-
-        $this->assertEquals(422, $response->getStatusCode());
-
-        $body = (string) $response->getBody();
-        $this->assertJson($body);
-        $this->assertEquals([
-            'errors' => [
-                [
-                    'status' => '422',
-                    'code' => 'validation_error',
-                    'detail' => 'The content field is required.',
-                    'source' => ['pointer' => '/data/attributes/content'],
-                ],
-            ],
-        ], json_decode($body, true));
-    }
-
-    /**
-     * @test
-     */
     public function cannot_create_discussion_without_title()
     {
         $response = $this->send(
